+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [10.2.0]
+## [10.3.0]
+
+### Added
+
+- txt
+  - new functions: bl64_txt_run_uniq, bl64_txt_run_sort
+- cnt
+  - new functions: bl64_cnt_login_stdin,bl64_cnt_container_is_running
+
+## [10.2.2]
 
 ### Added
 
@@ -225,7 +234,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - check
   - bl64_check_alert_failed
 
-[10.2.0]: https://github.com/serdigital64/bashlib64/compare/10.1.1...10.2.0
+[10.3.0]: https://github.com/serdigital64/bashlib64/compare/10.2.2...10.3.0
+[10.2.2]: https://github.com/serdigital64/bashlib64/compare/10.1.1...10.2.2
 [10.1.1]: https://github.com/serdigital64/bashlib64/compare/10.0.1...10.1.1
 [10.0.2]: https://github.com/serdigital64/bashlib64/compare/10.0.2...9.2.4
 [9.2.4]: https://github.com/serdigital64/bashlib64/compare/9.2.4...9.1.4